refactor: Changed file order
Moved field arrays into correct folder and other publication request form sections into their own respective folders
