+++ src/components/PublicationRequestFormSections/FieldArrays/index.js
@@ -4,4 +4,6 @@ export { default as FundingFieldArray } from './FundingFieldArray';
 export { default as IdentifiersFieldArray } from './IdentifiersFieldArray';
 export { default as OtherEmailFieldArray } from './OtherEmailFieldArray';
 export { default as StreetAddressesFieldArray } from './StreetAddressesFieldArray';
+export { default as PublicationStatusFieldArray } from './PublicationStatusFieldArray';
+
