--- src/components/PublicationRequestFormSections/PublicationStatusFieldArray/index.js
@@ -1,1 +1,0 @@
-export { default } from './PublicationStatusFieldArray';
