+++ src/components/PublicationRequestFormSections/PublicationStatusForm/index.js
@@ -1,0 +1,1 @@
+export { default } from './PublicationStatusForm';
